[WIP][DD-833] moved DB clean-ups tu persistence module GuicedTest
